[cleanup] move operations out of ent.ts
file is too big and it hasn't made sense in that file for a long long time
